tests: mutation: Fix test_external_memory_usage() to not measure mutation object footprint
The test measured copying of the mutation object, but verified the
measurement against mutation_partition::external_memory_usage(). So
anything allocated on the mutation object level would cause the test
to (incorrectly) fail. Fix that by copying only the mutation_partition
part.

Currently not a problem, because the partition_key is stored in the
in-line storage. Would become a problem once inline storage is
reduced.