api: gossiper: allow querying specific shard for alive/dead nodes
Introduce an optional "shard" parameter to `/gossiper/endpoint/live/`
and `/gossiper/endpoint/down/` endpoints which allows to query the
live/dead node set on a specific shard. If "shard" is not provided we
query shard 0 as before.

We'll use it in tests to wait until nodes mark other nodes as dead/alive
on every shard, e.g. before starting an operation which requires
some other nodes to be alive. Waiting on shard 0 only does not guarantee
that other shards have the same information yet (it takes a while for
this information to replicate between shards).