test: test_raft_upgrade: wait until alive nodes are marked as such before remove_node
In `test_raft_upgrade_with_node_remove` we restart 2 out of 3 nodes,
then remove the third node. This worked with non-RBNO removenode.
RBNO removenode fails because it requires other nodes to participate and
checks whether they are alive - but the test called the removenode
operation immediately, before gossiper managed to mark the other node
alive on each of the restarted node.

Fix this by waiting until the nodes mark each other as alive so they
only consider the dead node as dead. We must do this on every shard
because liveness information is not replicated instantaneously between
shards.

Ref: #12173