schema: operator<<: print also tombstone_gc_options
They are currently missing from the printout
when the a table is created, but they are determinal
to understanding the mode with which tombstones are to
be garbage-collected in the table.  gcGraceSeconds alone
is no longer enough since the introduction of
tombstone_gc_option in a8ad385ecd3e2b372db3c354492dbe57d9d91760.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>