test/cql-pytest: tests for null and empty in collections
This patch addes several tests which clarify issue #5825, which claimed
that Scylla may allow in certain cases insertion of null values into
a collection - which is supposed to be forbidden.

The tests demonstrates that it is an *empty* value (a value with length 0,
not a null with length -1), not a null value, which is what is allowed
inside a collection, and Python driver bug is what causes an attempt to write
a null value to result in an empty value. The latter is most visible with
a list<text> column - where what is written into the list element is an
empty string, not a null.

This driver bug has been long fixed in the Scylla fork of the branch
(https://github.com/scylladb/python-driver/issues/201), but was never
fixed in Datastax's (https://datastax-oss.atlassian.net/browse/PYTHON-1355)
so two of the new tests are skipped when using a Datastax driver (which
we do not do in our CI, but some developers may decided to do locally).

When the tests do try to write a null to a collection, they do report
an error expected.

Therefore, these tests show that issue #5825 is not a Scylla bug.
Two of the new tests fail on Cassandra (so they are marked cassandra_bug)
because it reports an internal server error "java.lang.NullPointerException"
instead of the expected InvalidRequest error.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>