test/cql-pytest: tests for null and empty in collections
This patch addes several tests which clarify issue #5825, which claimed
that Scylla may allow in certain cases insertion of null values into
a collection - which is supposed to be forbidden.

The tests demonstrates that it is an *empty* value (a value with length 0,
not a null with length -1), not a null value, which is what is allowed
inside a collection, and Python driver bug is what causes an attempt to write
a null value to result in an empty value. The latter is most visible with
a list<text> column - where what is written into the list lelement is an
empty string, not a null.

Therefore, these tests show that issue #5825 is not a Scylla bug. It can
be reported as a Python driver bug if we want - but note that if this bug
is fixed in the future, we'll need to change two of the tests.

Because the tests just verify the current behavior, they all pass on
Scylla. One test fails on Cassandra, which gets an internal server error
when an empty integer is written to a list.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>